refactor: asset 폴더 내 파일명 케밥 케이스로 수정
