+++ src/assets/profile-icon.png
